requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
anthropic==0.52.2
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
boto3==1.38.32
botocore==1.38.32
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
distro==1.9.0
dnspython==2.7.0
eval_type_backport==0.2.2
fasta2a==0.2.16
fastapi
fastavro==1.11.1
filelock==3.18.0
flask-cors==6.0.0
Flask==3.1.1
frozenlist==1.7.0
fuzzywuzzy==0.18.0
google-auth-oauthlib==1.2.2
google-auth==2.40.3
google-genai==1.19.0
griffe==1.7.3
groq==0.26.0
h11==0.14.0
hf-xet==1.1.3
httpcore==1.0.5
httpx-sse==0.4.0
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
logfire-api==3.18.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp
mistralai==1.8.1
motor
multidict==6.5.0
mysql-connector-python==9.3.0
oauthlib==3.2.2
openai==1.85.0
opentelemetry-api==1.34.0
outcome==1.3.0.post0
packaging==25.0
prompt_toolkit==3.0.51
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic-ai-slim==0.2.16
pydantic-ai==0.2.16
pydantic-evals==0.2.16
pydantic-graph==0.2.16
pydantic-settings==2.9.1
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
# python-Levenshtein==0.21.1
rapidfuzz==3.9.0
python-multipart==0.0.20
PyYAML==6.0.2
requests-oauthlib==2.0.0
requests==2.32.4
rfc3986==1.5.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
selenium>=4.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sse-starlette==2.3.6
starlette==0.46.2
tokenizers==0.21.1
tqdm==4.67.1
trio-websocket==0.12.2
trio==0.30.0
types-requests==2.32.0.20250602
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uuid==1.30
uvicorn[standard]
wcwidth==0.2.13
webdriver-manager>=3.8.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
youtube-search-python>=1.0.0
yt-dlp>=2023.01.01
fastapi
uvicorn[standard]==0.29.0
python-dotenv==1.1.0